refactor: InputContainer를 InfoInputSection에서 CardRegister/components로 이동
